Initial checkin of a new 'vfs' project which is based on Adam's current code base in jakarta-ant-myrmidon repository, in aut/src/java/org/apache/aut/vfs/.
There's a Maven build ready for use along with some extremely brief documentation. 
Right now we need to get the dependencies sorted out as it doesn't currently compile. 
There is a dependency on some excalibur and avalon code. Plus there's currently a dependency on ORO's FTP client. 
Maybe we should port this over to use the 'net' component in the sandbox (Apache's distro of this code).

Feel free to dive in and help!


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/vfs/trunk@147755 13f79535-47bb-0310-9956-ffa450edef68