pluggable files cache.
Till now VFS cached every file at least for the lifetime of the filesystem. But the filesystem lifes as long as the filesystem manager or until one calls close() (but who will do this?)
So VFS might tend to use more and more memory if using VFS as singleton (as proposed in the docs)

Currently only a NullFilesCache (no cache) and a DefaultFilesCache (current behaviour) exists.
But now there is a framework where we could begin to implement a more sophisticated cache.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/vfs/trunk@148016 13f79535-47bb-0310-9956-ffa450edef68