Now found the ultimative solution.
Introduce ResourceFileProvider.
This provider do not have its own filesystem implemention.
It simply resolve a full filename using the current or passed classloader and using the filesystemmanager to finally resolv the file.

In the case of a resource within an jar this will result in passing back a vfs jar-fileobject with all its goodies missing in an simple url connection.

The capability of such a provider is "Dispatcher" as it cant know much about the filesystem finally used to handle the resource.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/vfs/trunk@148043 13f79535-47bb-0310-9956-ffa450edef68