introduced FileContentInfo&FileContentInfoFactory
Now it is possible that a fileprovider can also provide informations about the content of a file.
Currently: content-type and content-encoding.

Currently only the Http filesystem is able to provide both. The values are the ones retrieved from the web-server.

All other filesystems are using the FileContentInfoFilenameFactory.
The content-type is determined by the extension of the filename using the URLConnection.getFileNameMap() jdk function.
The content-encoding is always null.

This default Factory can be override through the GlobalConfiguration.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/vfs/trunk@148048 13f79535-47bb-0310-9956-ffa450edef68