WebDav proxy handling
introduced a new (internal) way how to get the children of a folder.
Now the fileobject can return already resolved object (instead of simple strings) to e.g. prefill the file-type.

In WebDav you cant directly get the type of a resource when you do not have the rights on this folder, but the parent of the resource do know them.
Now if the parent do a "list children" those info is stored into the child by the parent.

Same story for the content-size, will fix this later.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/vfs/trunk@148052 13f79535-47bb-0310-9956-ffa450edef68