PR: 29469 Submitted By: rami.ojares -at- elisa.fi (Rami Ojares)
Not fully implemented yet (the main part is missing in this commit as we have to wait for the jsch-0.1.16 release)

Modified to honor the fact that sftp (jsch) cant set the lastmodtime on folders.
Split Capability.SET_LAST_MODIFICATION_TIME into
Capability.SET_LAST_MODIFICATION_TIME_FILE and
Capability.SET_LAST_MODIFICATION_TIME_FOLDER
Added a getLastModTimeAccuracy, mainly for the testcase-check.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/vfs/trunk@148099 13f79535-47bb-0310-9956-ffa450edef68