Again changed when and how to release resources. Now we count the number of "still accessible fileObject".
If, for some reason this do not work there now is also the FileSystemManager.closeFileSystem(FileSystem) method where you can explicitely close the fileSystem and free all its resources (including the cache).
This method might be dangerous if you use VFS as singleton (VFS.getManager()) and call it on a "global" fileSystem. This closes the filesystem for all other thread too.

git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/vfs/trunk@484648 13f79535-47bb-0310-9956-ffa450edef68