SFTP: Perform less strict permission checks on servers where the exec channel is disabled
Prior to this patch, all permissions checks would fail on servers
without an exec channel: SftpFileObject relied on the knowledge of the
current user's UID and GID. Neither of these can be obtained without an
exec channel. The SFTP specifications do not appear to mention any
alternative method of ascertaining the UID/GID.

As a workaround, when the UID can't be detected, SftpFileObject will use
a DummyPosixPermissions instance, which always "grants" RWX.

There are two sides to this coin:

1. This will allow normal operation against SFTP servers without an exec
channel.
2. This may cause JSchExceptions to be thrown when the actual
permissions are inconsistent with the dummy permissions.

This seems to be preferable to not being able to use exec-less SFTP
servers. Regressions should be limited to these kinds of servers.

Calling getUId when first making a connection might impose an overhead,
but this is presumably a trivial one.