1.fix bug for class not found:org.apache.commons.vfs2.provider.webdav.WebdavFileProvider since 2.5.0 this bug is caused by the remove of WebdavFileProvider, and when try to load WebdavFileProvider if the user does not have class WebdavFileProvider, then explode.
now the code will check the class as well as its dependency when load.(if the class itself does not exist then return)

2.add clean room test
This bug happened and even with this bug the tests now can pass.
that is because you include all things needed in test scope guys.
and we users do not usually include them all.
so when something went wrong and vfs can only run correctly when user code imports some repo(like this time), the current test mechanism can never tell.
So I add a clean room test for it.