SoftRefFilesCache: eliminate "volatile" on softRefReleaseThread
That "double checked locking" in startThread() is rather pointless,
because the overhead saved by not locking never materializes.
startThread() is only ever called when there is no thread, and
optimizing for a corner case with a data race isn't worth the
complexity.  So let's just do everything inside the lock and remove
"volatile".