SFTP: Memory leak because AbstractFileProvider#findFileSystem fails to detect equality of SFTP FileSystemOptions
When using FileSystemManager as a singleton and performing repeated SFTP operations, SftpFileSystem instances will accumulate with every resolveFile() invocation.  The cause for this is that AbstractOriginatingFileProvider#getFileSystem fails to locate the FileSystem from a previous invocation inside the TreeMap because all implementations of the SFTP IdentityProvider interface do not have equals() / hashCode() implementations and thus make calls to Arrays.deepHashcode() inside FileSystemOptions#compareTo always return different values, even if the FileSystemOption instances are actually equal.