Support URLencoding during normalization
Earlier, the path would be URL-decoded after normalization.
This meant some opportunities for normalization would be missed
when dots or slashes were URL-encoded.

One way to solve this would be to change the ordering of operations,
but that seems risky as it would change the meaning of some of the
abstractions. Because of that I opted for a more conservative
approach, where normalization will take into account URL-encoded
characters, but otherwise leave the input string intact.